Our own spaceship game.
Using non resetting velocities to replicate the lack of resistance in space. Power-ups exclusively add hull health (essentially life points) for now. Used "hull health" because I was ambitiously hoping to add a large ship in the background that we were protecting, but could just be implied. The core of this game works for sure. I added velocity status monitoring and the scores monitoring in the top. I tried to orient them in different places but had no luck for some reason. Maybe you'll know what to do.